move in the correct place
